21.07.2023 implementation Images Viewer

+++ app/src/main/java/com/example/shoplocalxml/ui/product_item/DividerItemRowDecoration.kt
@@ -44,7 +44,7 @@ class DividerItemRowDecoration : ItemDecoration() {
         for (i in 0 until childCount - 1) {
             val child = parent.getChildAt(i)
             val params = child.layoutParams as RecyclerView.LayoutParams
-            val top = (child.bottom + paddingProductCard/2f).toFloat()
+            val top = child.bottom + paddingProductCard/2f
             c.drawLine(start, top, end, top+1f, paint)
         }
     }
